CHORE : 리뷰 반영

+++ src/main/java/com/fakedevelopers/bidderbidder/dto/ProductInformationDto.java
@@ -16,7 +16,7 @@ public class ProductInformationDto {
     private final String category;
     private final long openingBid;
     private final Long hopePrice;
-    private final long tick;
+    private final int tick;
     private final String expirationDate;
     private final String createdDate;
     private final int bidderCount;
@@ -25,7 +25,7 @@ public class ProductInformationDto {
     private final List<BidDto> bids = new ArrayList<>();
 
     public ProductInformationDto(ProductEntity productEntity, List<String> images,
-        List<BidEntity> bidEntities) {
+            List<BidEntity> bidEntities) {
         this.productTitle = productEntity.getProductTitle();
         this.productContent = productEntity.getProductContent();
         this.category = productEntity.getCategory().getCategoryName();
@@ -33,9 +33,9 @@ public class ProductInformationDto {
         this.hopePrice = productEntity.getHopePrice();
         this.tick = productEntity.getTick();
         this.expirationDate = productEntity.getExpirationDate()
-            .format(DateTimeFormatter.ofPattern(Constants.DATE_FORMAT));
+                .format(DateTimeFormatter.ofPattern(Constants.DATE_FORMAT));
         this.createdDate = productEntity.getCreatedDate()
-            .format(DateTimeFormatter.ofPattern(Constants.DATE_FORMAT));
+                .format(DateTimeFormatter.ofPattern(Constants.DATE_FORMAT));
         this.bidderCount = bidEntities.size();
         this.images = images;
         int len = bidEntities.size();
