+++ src/main/java/com/fakedevelopers/bidderbidder/dto/ProductListDto.java
@@ -10,29 +10,30 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ProductListDto {
 
-  private final long productId;
-  private final String thumbnail;
-  private final String productTitle;
-  private final String category;
-  private final Long hopePrice;
-  private final long openingBid;
-  private final long tick;
-  private final String expirationDate;
-  private final int bidderCount;
+    private final long productId;
+    private final String thumbnail;
+    private final String productTitle;
+    private final String category;
+    private final Long hopePrice;
+    private final long openingBid;
+    private final int tick;
+    private final String expirationDate;
+    private final int bidderCount;
 
-  public ProductListDto(ProductEntity productEntity, boolean isWeb, Integer bidderCount) {
-    productId = productEntity.getProductId();
-    thumbnail =
-        "/product/getThumbnail?productId=" + productEntity.getProductId() + "&isWeb=" + isWeb;
-    productTitle = productEntity.getProductTitle();
-    category = productEntity.getCategory().getCategoryName();
-    hopePrice = productEntity.getHopePrice();
-    openingBid = productEntity.getOpeningBid();
-    tick = productEntity.getTick();
-    expirationDate =
-        productEntity
-            .getExpirationDate()
-            .format(DateTimeFormatter.ofPattern(Constants.DATE_FORMAT));
-    this.bidderCount = (bidderCount == null ? 0 : bidderCount);
-  }
+    public ProductListDto(ProductEntity productEntity, boolean isWeb, Integer bidderCount) {
+        productId = productEntity.getProductId();
+        thumbnail =
+                "/product/getThumbnail?productId=" + productEntity.getProductId() + "&isWeb="
+                        + isWeb;
+        productTitle = productEntity.getProductTitle();
+        category = productEntity.getCategory().getCategoryName();
+        hopePrice = productEntity.getHopePrice();
+        openingBid = productEntity.getOpeningBid();
+        tick = productEntity.getTick();
+        expirationDate =
+                productEntity
+                        .getExpirationDate()
+                        .format(DateTimeFormatter.ofPattern(Constants.DATE_FORMAT));
+        this.bidderCount = (bidderCount == null ? 0 : bidderCount);
+    }
 }
